pybcs:Fixed the bae uploading issue
Just used the default logging module of python instead of the
bae.app.logging.

This is not the best way to solve this problem. Since bae server
will redirect python logger to null so we couldn't get logs from
bae server.

I will keep this issue in mind and try to find another way to
solve this problem.

Signed-off-by: guofengliu <lgf612@sina.com>